requirement_extractor_service/requirements.txt
functions-framework
google-cloud-firestore
google-cloud-storage
vertexai
